// api/models/circuitCreateRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CircuitCreateRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  country: string;
  /** @minLength 1 */
  city: string;
  length: number;
  layoutUrl?: string;
  /** @minLength 1 */
  lapRecord: string;
  imageUrl?: string;
  firstParticipationYear: number;
  numberOfCorners: number;
  description?: string;
}